Update: BaseMountain 컴포넌트 분리 & dat.gui 제거
